+++ routes/user.route.js
@@ -1,0 +1,18 @@
+var express = require('express');
+
+var controller = require('../controllers/user.controller');
+
+var router = express.Router();
+
+router.get('/', controller.index);
+
+router.get('/search', controller.search);
+
+router.get('/create', controller.create);
+
+router.get('/:id', controller.get);
+
+router.post('/create', controller.post);
+
+
+module.exports = router;
