+++ views/index.pug
@@ -1,8 +1,10 @@
-h1 coders.tokyo
-a(href="/users/create") New User
-ul
-    each user in users
-        li
-            span= user.name
-            span &nbsp;
-            a(href="/users/" + user.id) View 
+extends ./layouts/layout.pug
+block content
+    h1 coders.tokyo
+    a(href="/users/create") New User
+    ul
+        each user in users
+            li
+                span= user.name
+                span &nbsp;
+                a(href="/users/" + user.id) View
