+++ views/users/create.pug
@@ -1,5 +1,7 @@
-h1 Create User 
-form(action="/users/create", method="POST")
-    label Name
-    input(name="name", type="text")
-    button Create
+extends ../layouts/layout.pug
+block content
+    h1 Create User 
+    form(action="/users/create", method="POST")
+        label Name
+        input(name="name", type="text")
+        button(class="btn btn-info ml-2") Create
