+++ views/users/index.pug
@@ -1,7 +1,9 @@
-h1 Users List 
-form(action="/users/search", method="GET") 
-    input(type="text", name="q", value= name)
-    button Search 
-ul 
-    each user in users
-        li= user.name
+extends ../layouts/layout.pug
+block content
+    h1 Users List 
+        form(action="/users/search", method="GET") 
+            input(type="text", name="q", value= name)
+            button Search 
+        ul 
+            each user in users
+                li= user.name
